mild refactor, single file for agents
