+++ src/scripts/example_agent_run.py
@@ -1,8 +1,8 @@
-from benchmark.agent.types import AgentConfig
-from benchmark.agent.hypothesizer import PythonAgent
-from benchmark.agent.sorryer import LeanAgent
-from scripts.config import leancfg, pythoncfg
 import pathlib
+
+from benchmark.agent.agents import LeanAgent, PythonAgent
+from benchmark.agent.types import AgentConfig
+from scripts.config import leancfg, pythoncfg
 
 EXAMPLE = "string"
